Split login into activity and fragments.
