+++ app/src/main/res/layout/activity_login.xml
@@ -22,17 +22,6 @@
             android:contentDescription="@string/content_description_image_placeholder"
             android:src="@drawable/logo_white_background" />
 
-<!--        <TextView
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="@string/app_title"
-            android:textColor="@color/theme_secondary_text_color"
-            android:textSize="@dimen/title_text_size"
-            android:layout_centerVertical="true"
-            android:layout_toEndOf="@+id/app_image"
-            android:layout_marginStart="30dp"
-            android:layout_marginLeft="30dp" />-->
-
     </RelativeLayout>
 
     <LinearLayout
@@ -108,11 +97,23 @@
     </LinearLayout>
 
     <TextView
-        android:id="@+id/textView"
+        android:id="@+id/register_text"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
         android:layout_alignParentStart="true"
         android:layout_below="@+id/email_login_form"
+        android:layout_marginTop="10dp"
+        android:gravity="center"
+        android:text="@string/sign_up"
+        android:textSize="@dimen/login_text_size"
+        android:textColor="@color/theme_secondary_text_color" />
+
+    <TextView
+        android:id="@+id/forgot_password"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:layout_alignParentStart="true"
+        android:layout_below="@+id/register_text"
         android:layout_marginTop="10dp"
         android:gravity="center"
         android:text="@string/forgot_password"
