+++ app/src/main/res/values/strings.xml
@@ -130,7 +130,7 @@
     <!-- Strings related to login -->
     <string name="prompt_email">Email</string>
     <string name="prompt_password">Password</string>
-    <string name="action_sign_in">Sign in or register</string>
+    <string name="action_sign_in">Sign in</string>
     <string name="action_sign_in_short">Sign in</string>
     <string name="connect_to_network_message">Unable to connect to network</string>
     <string name="go_to_settings_message">Go to settings</string>
@@ -140,6 +140,7 @@
     <string name="login_text">Login</string>
     <string name="error_playing_song_message">There was a problem playing this song.</string>
     <string name="forgot_password">Forgot Password ?</string>
+    <string name="sign_up">Sign Up</string>
 
     <!-- Item updated messages -->
     <string name="item_updated_mix">This mix has been updated.</string>
